Move SwapChainRenderTarget to Platform specific
